Add saving of option to disable traffic lights
The option is saved in SaveGames and isn't global.

Use word Level instead of Game. Rename MainMachine to Level.
Level is used by C:S: OnLevelLoaded, OnLevelUnloading, etc..

Extract C:S mod interfaces into IUserMod class.
As a result only one class is created for all C:S mod interface classes.

Remove unused classes.
