+++ src/ToggleTrafficLights/Game/Options.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 namespace Craxy.CitiesSkylines.ToggleTrafficLights.Game
 {
-  public class Options
+  public sealed class Options
   {
     public const string SettingsFile = "ToggleTrafficLights";
 
@@ -65,16 +65,22 @@ namespace Craxy.CitiesSkylines.ToggleTrafficLights.Game
     {
       throw new NotImplementedException();
     }
-    
-    #region settings on savegame basis
+  }
+
+  public sealed class SaveGameOptions
+  {
     public enum TrafficLightsHandling : byte
     {
       Default,
       NoTrafficLights,
       AllTrafficLights
     }
-    public TrafficLightsHandling TrafficLights { get; set; } = TrafficLightsHandling.Default;
-    
-    #endregion settings on savegame basis
+    private static readonly TrafficLightsHandling DefaultTrafficLights = TrafficLightsHandling.Default;
+    public TrafficLightsHandling TrafficLights { get; set; } = DefaultTrafficLights;
+
+    public void Reset()
+    {
+      TrafficLights = DefaultTrafficLights;
+    }
   }
 }
